Provisioning (#259)
* First stab at provisioning

* Added support for backend provisioning

* Add provisioned flag to Backend / Tool

* Prevent modification of provisioned backends

* Prevent deletion of provisioned backend

* Prevent modification / deletion of a provisioned tool

* Protect against malicious API usage by admin

* Expand env variables in openapi tool auth headers

* Enabled expansion only for provisioned tool

* env expansion for backend api key

* Provisioning from directory

* Renamed provision location env var

* Awaiting... god save linter

+++ logicle/app/api/tools/[toolId]/route.ts
@@ -20,10 +20,18 @@ export const GET = requireAdmin(async (req: Request, params: { toolId: string })
 
 export const PATCH = requireAdmin(async (req: Request, params: { toolId: string }) => {
   const data = await req.json()
+  const existingTool = await getTool(params.toolId)
+  if (!existingTool) {
+    return ApiResponses.noSuchEntity('No such backend')
+  }
+  if (existingTool.provisioned) {
+    return ApiResponses.forbiddenAction("Can't modify a provisioned tool")
+  }
   await updateTool(params.toolId, {
     ...data,
     id: undefined,
     type: undefined,
+    provisioned: undefined, // protect against malicious API usage
     createdAt: undefined,
     updatedAt: new Date().toISOString(),
   })
@@ -31,6 +39,13 @@ export const PATCH = requireAdmin(async (req: Request, params: { toolId: string 
 })
 
 export const DELETE = requireAdmin(async (req: Request, params: { toolId: string }) => {
+  const existingTool = await getTool(params.toolId)
+  if (!existingTool) {
+    return ApiResponses.noSuchEntity('No such backend')
+  }
+  if (existingTool.provisioned) {
+    return ApiResponses.forbiddenAction("Can't delete a provisioned tool")
+  }
   try {
     await deleteTool(params.toolId)
   } catch (e) {
